WPP99 - SNI support with httpclient lib
Upgrade of the HttpClient from 4.2 to 4.5.2. This is solving SNI bug problems and HttpClient can be more configurable.

Just a warning the current use of the httpclient can be made on a better way with this new version and will improve the configuration possibilities.
Also the configuration by default use a PooledConnection, just need some customization.